Correção da função ft_prinf.c

+++ ft_printf/ft_print_u.c
@@ -15,7 +15,7 @@
 void	ft_print_u(t_ptf *parms)
 {
     parms->v_p = va_arg(parms->ap, void *);
-    parms->v_str = ft_itoa_base((unsigned int)parms->v_p, DEC_BASE);
+    parms->v_str = ft_itoa_base((uintmax_t)parms->v_p, DEC_BASE);
     parms->len_c = ft_strlen(parms->v_str);
 	parms->len += ft_strlen(parms->v_str);
 }
